Adiciona suporte à consulta de contas passivas
- Criação do formulário `ConsultadorContasPassivas` para gerenciar e consultar contas passivas.
- Adicionados novos controles ao formulário principal (`MainTab`) para exibir e buscar informações de contas passivas.
- Implementados métodos `GetContasPassivas` e `GetContasAtivas` em `DBConfig` para acesso ao banco de dados.
- Alterado o tipo de `numConta` para `long` nas classes `ContaPassiva` e `ContaAtiva`.
- Adicionado o método `ConsultarContasPassivas` em `Program` para abrir o novo formulário.
- Ajustes gerais na interface do usuário e melhorias no tratamento de erros.

+++ Scripts/GerenciadorObjetos.cs
@@ -15,13 +15,13 @@ namespace ImportadorContaMovimentacao.Scripts
 
     public class ContaPassiva
     {
-        public int numConta { get; set; }
+        public long numConta { get; set; }
         public string? nomeConta { get; set; }
         public string? contaAnalitica { get; set; }
     }
     public class ContaAtiva
     {
-        public int numConta { get; set; }
+        public long numConta { get; set; }
         public string? nomeConta { get; set; }
         public string? contaAnalitica { get; set; }
     }
